Change labels and number type for the 'U' column in the tool table.
Change 'U offsets' to Max RPM
Change from float to int
Center Align the Values
Allow -1 to 24000 for RPM values.

+++ src/qtpyvcp/lib/db_tool/tool_table.py
@@ -34,7 +34,7 @@ class Tool(Base):
     i_offset = Column(Float)
     j_offset = Column(Float)
     q_offset = Column(Float)
-    u_offset = Column(Float)
+    u_offset = Column(Integer)
     v_offset = Column(Float)
     w_offset = Column(Float)
     diameter = Column(Float)
